email_genai copy/requirements.txt

streamlit
requests
beautifulsoup4
langchain-core
langchain-groq
python-dotenv
urllib3